Encoded video (#5)
Summary:
Pull Request resolved: https://github.com/fairinternal/pytorchvideo/pull/5

Adds an encoded video class that abstracts away decoding and clip extraction for videos. I haven’t been able to test the case where a video doesn’t have the header information, does anyone have an example of a video like that?

I used pyav to start with because it was faster and slightly more accurate in my benchmarks compared to torchvision. Will switch it to torchvision when that’s resolved

Next diffs:
- audio support
- kinetics dataloader
- benchmarking
- add support for multiple backends (pyav/torchvision/dicord)
- add metadata caching support (need a dataset to benchmark it on)
- Multi-clip decoding for performance

Reviewed By: takatosp1, lyttonhao

Differential Revision: D22478582

fbshipit-source-id: 68085e816cbf86c1b611d3232afbfefe48185f8e