Fix kinetics test
Summary: There has been a strange flaky bug in the kinetics test where buck fails to fetch the results when in mode/opt mode. I've worked out that the cause is to do with an interaction between torchvision and pyav. I'm not exactly sure why this is the case, but the changes in the diff appear to fix everything consistently.

Reviewed By: takatosp1

Differential Revision: D22960584

fbshipit-source-id: 46e232f8406362516aa2b64ad6120563a3cfcda5