EpicKithcen Encoded Video Dataset
Summary:
This is a functional implementation of the epic kitchen dataset that now can load from encoded videos and frame videos. It was quite simple with the AbstractVideo abstraction. I figured I'd show what it looks like.

I am pausing here to refactor the underlying dataset to be iterable rather than mapped.

Reviewed By: tullie, takatosp1

Differential Revision: D23409559

fbshipit-source-id: d05311eba115617ef50b72b9b62793d15271c5df