Fix failing test on github due to dependency on ls undocumented behavior
Summary:
Currently this test depends on PathManager.ls returning files in frame-order. This occurs on local and FB test machines as the files are written in order. This does not happen on the machines where the github repo tests this code.

Relying on this assumption of ls' behavior is not great in the first place. Here we sort the frames first before processesing to ensure that they are in order.

We leave the previous failing assert as we want to be able to fail hard when there are missing files within a video. Handling missing files is up to the user.

Reviewed By: tullie

Differential Revision: D24688753

fbshipit-source-id: 4dd77397d97534b71c34067a3ac7d40fed350cec