FAIM model masked components
Summary:
The FAIM video model code consists of the following components:

* Temporal_convolution (conv/norm/relu/)
* MLP
* Transformer
* Attention_multihead
* Lstm
* Gate_layer
* Learn_missing_default
* Squeeze_excitation_gate
* Embedding_non_local_gate
* Positional_encoding
* Temporal_pooling
* Attention_fc_softmax
* Normalize
* Embedding
* FusionLayer

Then there is a skeleton that puts these together as a multi-stream fusion architecture.

Of the components the FAIM 2020a model uses:
* MLP
* Attention_multihead
* FusionLayer
* Normalize
* Positional_encoding
* Temporal_pooling

This diff converts these FAIM 2020a components to PTV style for use with a PTV MultiStreamFuse skeleton. Using the skeleton with these components the 2020a model can be constructed with a F6-Hydra or Classy-Json wrapper.

The next few diffs:
* Add flat interface which composes components
* Classy Vision wrapper
* Adding the rest of the FAIM Video components

See design doc for more info:
https://docs.google.com/document/d/1nz7YEihhGniuKP7ZLs5n7dnwZCWCWdDYPZ-Lcn1OpmI/edit#heading=h.i0bpqfsp8omh

Reviewed By: takatosp1, HengCV

Differential Revision: D24278433

fbshipit-source-id: 83775e95b36f1f73d2bf0440b0408ca375623727