Fix RandomSampler
Summary:
This diff does two things:
1. Fixes the RandomSampler to avoid duplicating values when using multiple PyTorch DataLoader processes
2. Refactors the tests to use a mock_csv function

I'm not that happy with the solution for 1., having to condition on the type of sampler makes me think we need a better way of doing this. Potentially we could just have an enum that indicates a video sampler strategy. The problem with this is it's hard to generalize to arbitrary distributed cases.

Reviewed By: takatosp1

Differential Revision: D25369593

fbshipit-source-id: a2189be750b9129dd4afc55eb973d1aa6f2b96c0