Refactor create_slowfast to make the Net creation more clear
Summary:
The existing code puts all the stages into blocks, which are then split into pathway stages and fusion stages at the end.

The problem is that this assumes we always alternate pathway and fusion stages, which isn't necessarily true. Instead build the MultiPathwayWithFusion for each stage.

This allows us to remove the 1 item ModuleList near the end, and one PoolConcatPathway

Reviewed By: takatosp1

Differential Revision: D25809919

fbshipit-source-id: 41733ceb801dd1ab907e0b125411d0780153dadb