Fix acoustic resnet and benchmark (#35)
Summary:
Fix acoustic resnet and benchmark it on the video classification task. I was able to get 0.175 accuracy with this resnet implementation (the previous implementation I couldn't get more than 0.02 acc). Note that the AVSlowFast seems to suggest that the acoustic resnet on Kinetics should get about ~0.2 accuracy. I think this could be achieved with a bit more parameter tuning, particularly related to optimization parameters, however, this is overkill for the example.

Pull Request resolved: https://github.com/fairinternal/pytorchvideo/pull/35

Reviewed By: bxiong1202

Differential Revision: D27945328

Pulled By: tullie

fbshipit-source-id: 9719a86d4a30e545d648273b1963dc3ec7f2376a