Fix acoustic resnet and benchmark (#33)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytorchvideo/pull/33

Fix acoustic resnet and benchmark it on the video classification task. I was able to get 0.175 accuracy with this resnet implementation (the previous implementation I couldn't get more than 0.02 acc). Note that the AVSlowFast seems to suggest that the acoustic resnet on Kinetics should get about ~0.2 accuracy. I think this could be achieved with a bit more parameter tuning, particularly related to optimization parameters, however, this is overkill for the example.

Pull Request resolved: https://github.com/fairinternal/pytorchvideo/pull/35

Reviewed By: bxiong1202

Differential Revision: D27945328

Pulled By: tullie

fbshipit-source-id: 216e7d67e6ec92572c09b8a495822adbbf5a6158