PyTorchVideo + PyTorch3D demo (#44)
Summary:
## Motivation and Context

A demo which uses the PyTorchVideo `EncodedVideo` class to load and decode a video and train a NeRF model.

This project contains the dataloader, config file, utils files and instructions to download a video from the Objectron dataset and train a NeRF model using the implementation from PyTorch3D.

## How Has This Been Tested

Manually tested all the instructions and trained a model.

## Types of changes

A new project/tutorial.

## Checklist

- [x ] My code follows the code style of this project.
- [- ] My change requires a change to the documentation.
- [x ] I have updated the documentation accordingly.
- [x ] I have read the **CONTRIBUTING** document.
- [x ] I have completed my CLA (see **CONTRIBUTING**)
- [- ] I have added tests to cover my changes.
- [ x] All new and existing tests passed.

Pull Request resolved: https://github.com/fairinternal/pytorchvideo/pull/44

Reviewed By: lyttonhao

Differential Revision: D28748145

Pulled By: nikhilaravi

fbshipit-source-id: d2da3f8d383b6a448665b97fb01996883798625f