make MViT class interchangable in create_multiscale_vision_transformers
Summary:
in PostRay MViT use case, we will need to modify the forward logic in MViT a bit, but still would like to use the create_multiscale_vision_transformers function without duplicating a lot of codes.

This diff is to make the mvit class as a callable arg in create_multiscale_vision_transformers.

Differential Revision: D34900946

fbshipit-source-id: 342453d32432e245388a89166cc1bbe295bfd179