video MAE model
Summary:
Based on image MAE model, Haoqi has developed this video MAE model. Here we try to migrate the work Haoqi did in our internal library to OSS.

This diff is what will be submitted to OSS. As you can see, the masked_autoencoder.py and vit.py files only depends on layers/attention.py and layers/positional_encoding.py, and layers/attention.py and layers/positional_encoding.py only depend on standard library.
D35935368 (depending on D35935043), setup the environment for running training / eval. As you can see in this diff, the 4 files mentioned in D35935043 are not changed. With this setting, we run train/eval with D35935368 to validate the model migration. We will compare the top1 accuracy and top5 accuracy result from this diff and compare with previous runs. Once pass, we only need to submit D35935043

Reviewed By: haooooooqi

Differential Revision: D35935043

fbshipit-source-id: 6d5cf0b1f2b17d21cdc317d2a4c344082071b332