make MViT class interchangable in create_multiscale_vision_transformers (#50)
Summary:
X-link: https://github.com/fairinternal/pytorchvideo/pull/50

Pull Request resolved: https://github.com/facebookresearch/pytorchvideo/pull/179

in PostRay MViT use case, we will need to modify the forward logic in MViT a bit, but still would like to use the create_multiscale_vision_transformers function without duplicating a lot of codes.

This diff is to make the mvit class as a callable arg in create_multiscale_vision_transformers.

Reviewed By: feichtenhofer

Differential Revision: D34900946

fbshipit-source-id: a3a893c8a0b7bbe9592ea1bb3c27849c5ec84331