Stop using the github target directly
Summary:
I just search+replaced in VSCode to start using the third-party/pypi/... path, then I ran `arc f` to fix lints.

The goal here is to delete the first repo from the fbcode import tool by migrating it to mgt. This is a test of the process that I'm going to tell everyone else to go do (probably by filing tasks)

drop-conflicts

Reviewed By: jurajh-fb

Differential Revision: D40400653

fbshipit-source-id: f0cb54d3a9870a2422a79469f2143d3d17a23198